--- app/controllers/training_controller.rb
@@ -1,16 +1,0 @@
-class TrainingController < ApplicationController
-  def index
-  end
-
-  def show
-  end
-
-  def edit
-  end
-
-  def my_list
-  end
-
-  def new
-  end
-end
