+++ app/helpers/practices_helper.rb
@@ -1,0 +1,2 @@
+module PracticesHelper
+end
