--- app/helpers/training_helper.rb
@@ -1,2 +1,0 @@
-module TrainingHelper
-end
