+++ app/models/practice.rb
@@ -1,2 +1,3 @@
 class Practice < ApplicationRecord
+  belongs_to :user
 end
