+++ app/models/user.rb
@@ -13,6 +13,7 @@ class User < ApplicationRecord
 
   has_many :songs, dependent: :destroy
   has_many :users, dependent: :destroy
+  has_many :practices, dependent: :destroy
   has_many :comments
   has_many :likes
   has_many :liked_songs, through: :likes, source: :song
