--- app/views/training/edit.html.erb
@@ -1,2 +1,0 @@
-<h1>Training#edit</h1>
-<p>Find me in app/views/training/edit.html.erb</p>
