--- app/views/training/index.html.erb
@@ -1,2 +1,0 @@
-<h1>Training#index</h1>
-<p>Find me in app/views/training/index.html.erb</p>
