--- app/views/training/my_list.html.erb
@@ -1,2 +1,0 @@
-<h1>Training#my_list</h1>
-<p>Find me in app/views/training/my_list.html.erb</p>
