--- app/views/training/new.html.erb
@@ -1,2 +1,0 @@
-<h1>Training#new</h1>
-<p>Find me in app/views/training/new.html.erb</p>
