--- app/views/training/show.html.erb
@@ -1,2 +1,0 @@
-<h1>Training#show</h1>
-<p>Find me in app/views/training/show.html.erb</p>
