+++ spec/controllers/practices_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe PracticesController, type: :controller do
+
+end
